Add files via upload
The heart rate, or pace of the heartbeat, is measured using a device called a heartbeat sensor. The fundamental actions we do to maintain our health include checking our body's temperature, heart rate, and blood pressure.

We use thermometers to gauge body temperature and a sphygmomanometer to track blood pressure or arterial pressure.

There are two ways to check your heart rate: manually by feeling your wrists or neck for a pulse, or by using a heartbeat sensor.